+++ python/hooks/default.nix
@@ -1,6 +1,14 @@
-{ makeSetupHook, writeShellScriptBin, runCommandLocal, python310, bat, findutils, lib }:
+{ makeSetupHook, writeShellScriptBin, python310, bat, findutils, lib }:
 let
-  mergeConfigs = src: name: key: files:
+  generateConfigurationRunner =
+    { toolDerivation
+    , key
+    , config
+    , toolName ? toolDerivation.pname or toolDerivation.name
+    , configFlag ? "--config"
+    , extraArgs ? ""
+    , files ? [ ]
+    }:
     let
       remove = builtins.foldl' (acc: cur: "${acc} ${cur}") ""
         (builtins.map (file: "${file.key}:${file.path}=${file.removeField}")
@@ -8,57 +16,60 @@ let
             (file: builtins.isAttrs file && file ? removeField)
             files));
     in
-    runCommandLocal name { } ''
-      export PYTHONPATH=''${PYTHONPATH:-}:${python310.pkgs.toml}/${python310.sitePackages}
-      ${python310}/bin/python ${./config-merger.py} --tool "${key}" ${if remove != "" then "--remove-fields ${remove}" else ""} --files ${builtins.concatStringsSep " " (
-        builtins.map (file:
-          if builtins.isPath file.path then
-            "${file.path}=${file.key}"
-          else
-            "${src}/${file.path}=${file.key}"
-        )
-        (
-          builtins.map
-            (path: if builtins.isAttrs path then path else {inherit key path;})
-            files
-        )
-      )}
-    '';
-
-  generateConfigurationRunner =
-    { toolDerivation
-    , toolName ? toolDerivation.pname or toolDerivation.name
-    , configFlag ? "--config"
-    , config
-    , extraArgs ? ""
-    }:
     writeShellScriptBin toolName ''
+      config_file=$TMP/lint-configs/${config}
+      mkdir -p "$(dirname "$config_file")"
+      export PYTHONPATH=''${PYTHONPATH:-}:${python310.pkgs.toml}/${python310.sitePackages}
+      ${python310}/bin/python \
+        ${./config-merger.py} \
+        --tool "${key}" \
+        ${if remove != "" then "--remove-fields ${remove}" else ""} \
+        --files ${builtins.concatStringsSep " " (
+            builtins.map (file:
+              if builtins.isPath file.path then
+                "${file.path}=${file.key}"
+              else
+                "./${file.path}=${file.key}"
+            )
+            (
+              builtins.map
+                (path: if builtins.isAttrs path then path else {inherit key path;})
+                files
+            )
+        )} \
+        --out-file="$config_file"
+
+
       if [[ $@ =~ "--print-generated-config-path" ]]; then
-        echo "${config}"
+        echo "$config_file"
         exit 0
       fi
 
       if [[ $@ =~ "--print-generated-config" ]]; then
-        ${bat}/bin/bat "${config}"
+        ${bat}/bin/bat "$config_file"
         exit 0
       fi
 
-      ${toolDerivation}/bin/${toolName} ${configFlag} ${config} "$@" ${extraArgs}
+      ${toolDerivation}/bin/${toolName} ${configFlag} "$config_file" "$@" ${extraArgs}
     '';
 
-  blackWithConfig = src: toolDerivation: generateConfigurationRunner {
+  blackWithConfig = toolDerivation: generateConfigurationRunner {
     inherit toolDerivation;
-    config = mergeConfigs src "black.toml" "black" [
+    key = "black";
+    config = "black.toml";
+    files = [
       { path = "pyproject.toml"; key = "tool.black"; }
       "setup.cfg"
       { path = ./config/black.toml; key = "tool.black"; }
     ];
   };
 
-  coverageWithConfig = src: toolDerivation: generateConfigurationRunner {
+  coverageWithConfig = toolDerivation: generateConfigurationRunner {
     inherit toolDerivation;
+    key = "coverage";
+    config = "coverage.toml";
     configFlag = "--rcfile";
-    config = mergeConfigs src "coverage.toml" "coverage" [
+    files = [
       ".coveragerc"
       { path = "setup.cfg"; key = "tool:coverage"; }
       { path = "tox.ini"; key = "tool:coverage"; }
@@ -67,9 +78,11 @@ let
     ];
   };
 
-  flake8WithConfig = src: toolDerivation: generateConfigurationRunner {
+  flake8WithConfig = toolDerivation: generateConfigurationRunner {
     inherit toolDerivation;
-    config = mergeConfigs src "flake8.ini" "flake8" [
+    key = "flake8";
+    config = "flake8.ini";
+    files = [
       ".flake8"
       "setup.cfg"
       "tox.ini"
@@ -78,11 +91,13 @@ let
     ];
   };
 
-  isortWithConfig = src: toolDerivation: generateConfigurationRunner {
+  isortWithConfig = toolDerivation: generateConfigurationRunner {
     inherit toolDerivation;
     configFlag = "--settings-file";
     extraArgs = "--src-path .";
-    config = mergeConfigs src "isort.ini" "isort" [
+    key = "isort";
+    config = "isort.ini";
+    files = [
       ".isort.cfg"
       { path = "pyproject.toml"; key = "tool.isort"; }
       "setup.cfg"
@@ -92,10 +107,12 @@ let
     ];
   };
 
-  mypyWithConfig = src: toolDerivation: generateConfigurationRunner {
+  mypyWithConfig = toolDerivation: generateConfigurationRunner {
     inherit toolDerivation;
     configFlag = "--config-file";
-    config = mergeConfigs src "mypy.ini" "mypy" [
+    key = "mypy";
+    config = "mypy.ini";
+    files = [
       "mypy.ini"
       ".mypy.ini"
       { path = "pyproject.toml"; key = "tool.mypy.overrides"; }
@@ -105,10 +122,12 @@ let
     ];
   };
 
-  pylintWithConfig = src: toolDerivation: generateConfigurationRunner {
+  pylintWithConfig = toolDerivation: generateConfigurationRunner {
     inherit toolDerivation;
     configFlag = "--rcfile";
-    config = mergeConfigs src "pylint.toml" "pylint" [
+    config = "pylint.toml";
+    key = "pylint";
+    files = [
       { path = "pylintrc"; key = ""; }
       { path = ".pylintrc"; key = ""; }
       { path = "pyproject.toml"; key = "tool.pylint"; }
@@ -123,10 +142,12 @@ let
     ];
   };
 
-  pytestWithConfig = src: toolDerivation: generateConfigurationRunner {
+  pytestWithConfig = toolDerivation: generateConfigurationRunner {
     inherit toolDerivation;
     configFlag = "-c";
-    config = mergeConfigs src "pytest.ini" "pytest" [
+    config = "pytest.ini";
+    key = "pytest";
+    files = [
       "pytest.ini"
       { path = "pyproject.toml"; key = "tool.pytest.ini_options"; }
       { path = "pylintrc.toml"; key = "tool.pytest.ini_options"; }
@@ -138,18 +159,18 @@ let
   };
 in
 {
-  check = src: pythonPkgs:
+  check = _: pythonPkgs:
     makeSetupHook
       {
         name = "check-hook";
         deps = with pythonPkgs; [
           findutils
-          (coverageWithConfig src coverage)
-          (flake8WithConfig src flake8)
-          (isortWithConfig src isort)
-          (mypyWithConfig src mypy)
-          (pylintWithConfig src pylint)
-          (pytestWithConfig src pytest)
+          (coverageWithConfig coverage)
+          (flake8WithConfig flake8)
+          (isortWithConfig isort)
+          (mypyWithConfig mypy)
+          (pylintWithConfig pylint)
+          (pytestWithConfig pytest)
           # pytest is also useful as a module in PYTHONPATH for fixtures and such
           pytest
         ]
@@ -157,6 +178,6 @@ in
         lib.optional
           (with pythonPkgs.python.stdenv;
           !(isAarch64 && isDarwin && lib.versionOlder lib.version "22.11pre-git"))
-          (blackWithConfig src black);
+          (blackWithConfig black);
       } ./check.bash;
 }