+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Terraform deploy will now exit with non-zero exit code if deploy failed.
 
+### Changed
+- Python linting tools now combines their configs during runtime instead of build time.
+
 ## [1.2.2] - 2023-06-9
 
 ### Fixed